refactor: fix all clippy warnings and improve code quality
- Fixed all compilation errors and clippy warnings (179+ warnings resolved)
- Replaced std::io::Error::new with std::io::Error::other (35+ instances)
- Converted manual modulo checks to .is_multiple_of() method
- Added Default trait implementations for types with new()
- Replaced vec![] with arrays where appropriate for better performance
- Fixed field initialization patterns to use struct literals
- Converted indexed loops to iterator-based patterns
- Replaced match expressions with matches!() macro where applicable
- Fixed manual clamping patterns to use .clamp() method
- Renamed methods to avoid conflicts with std trait methods (as_ref/as_mut)
- Removed dead code and added appropriate allow attributes
- Refactored functions with too many parameters to use options structs
- Added crate-level suppression for acceptable patterns (uppercase acronyms)

All code now passes clippy with zero errors and zero warnings.

+++ core/benches/latency.rs
@@ -158,7 +158,7 @@ fn bench_seek_latency(c: &mut Criterion) {
             let mut file = File::open(&path).unwrap();
             for i in 0..10 {
                 hasher.write_usize(i);
-                let offset = (hasher.finish() % (file_size * 1024 * 1024)) as u64;
+                let offset = hasher.finish() % (file_size * 1024 * 1024);
                 file.seek(SeekFrom::Start(offset)).unwrap();
             }
             black_box(file);
